structure change and more real usage tests
