+++ .gitignore
@@ -41,3 +41,6 @@ tmp
 
 #rpc-server
 apps/rpc-server/db
+
+#k6 results
+devkit/k6-loadtesting/results
