+++ devkit/k6-loadtesting/README.md
@@ -1,0 +1,12 @@
+# K6 Loadtesting
+
+## Installation 
+
+This package utilizes k6 by Grafana Labs.
+Install 6 on test machine: https://k6.io/docs/get-started/installation/
+
+
+## How to test RPC Server
+
+1. Spin up Sandbox 
+2. Spin up RPC Server on http://localhost:3040/
